# Client: Toggle songs by pack instead of per song