--- a/DataHandler.py
+++ b/DataHandler.py
@@ -189,8 +189,7 @@ def extract_mod_data_to_json() -> list[Any]:
     """
 
     user_path = Utils.user_path(Utils.get_settings()["generator"]["player_files_path"])
-    folder_path = sys.argv[
-        sys.argv.index("--player_files_path") + 1] if "--player_files_path" in sys.argv else user_path
+    folder_path = sys.argv[sys.argv.index("--player_files_path") + 1] if "--player_files_path" in sys.argv else user_path
 
     print(f"Checking YAMLs for megamix_mod_data at {folder_path}")