Currently on connect the Client iterates through received network items and toggles songs one at a time and all that entails (open, scan, modify, write, repeat per individual item).
There have been recent reports of the Client timing out on connect, resulting in an infinite loop and a bricked Client.

This PR changes the song toggle logic to first reverse the incoming item IDs back to song IDs and their related song packs.
Per song pack:
- Open the mod_pv_db.txt once
- Per song ID, modify as needed
- Write out the mod_pv_db.txt once

From my early testing this results in around a 9x speedup in the time spent toggling, but could vary.
To test I hosted a solo room then ran `/collect Player` so every available song would need to be toggled.
The average worst case 250 base+DLC all songs (-1 for goal), 500 locations, went from 16.5 seconds to 2 seconds (on my machine).
An extreme worst case 1253 songs, 2506 locations, went from 84 seconds to 8.5 seconds (on my machine).

The toggle process is actually ran twice on connect. On the connect, obviously, but again when the DataPackage is received where it resyncs. This may need to be looked into.

To-do:
 - [ ] Refactor shared song_unlock preprocessing code in Client.py
 - [x] Make to-do list